+++ lib/base/base_bloc_page_state.dart
@@ -52,17 +52,29 @@ import 'base_page_delegate.dart';
 /// ```
 abstract class BaseBlocPageState<T extends StatefulWidget, B extends MainBloc>
     extends BasePageDelegate<T, B> {
+  void hideLoading({String? key = LoadingKey.global}) {
+    commonBloc.add(
+        SetComponentLoading(key: key ?? LoadingKey.global, isLoading: false));
+  }
+
   @override
   Widget buildLoadingOverlay({
     required Widget child,
     String? loadingKey = LoadingKey.global,
     Widget? loadingWidget,
+    Duration timeout = const Duration(seconds: 30),
   }) {
     return BlocBuilder<CommonBloc, CommonState>(
       buildWhen: (previous, current) =>
           previous.isLoading(key: loadingKey) !=
           current.isLoading(key: loadingKey),
       builder: (context, state) {
+        if (state.isLoading(key: loadingKey)) {
+          Future.delayed(timeout, () {
+            if (mounted && state.isLoading(key: loadingKey))
+              hideLoading(key: loadingKey);
+          });
+        }
         return Stack(
           children: [
             child,