+++ lib/base/base_page_delegate.dart
@@ -59,6 +59,18 @@ abstract class BasePageDelegate<T extends StatefulWidget,
     ..commonBloc = commonBloc
     ..navigator = navigator;
 
+  final Map<String, Widget> _widgetCache = {};
+
+  Widget getCachedWidget(String key, Widget Function() builder) {
+    return _widgetCache.putIfAbsent(key, builder);
+  }
+
+  @override
+  void dispose() {
+    _widgetCache.clear();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(